refactor : user model-domain 분리
